add authorization logic for destroy action

+++ app/controllers/api/v1/encounters_controller.rb
@@ -31,6 +31,7 @@ module Api
 
       def destroy
         encounter = Encounter.find(params[:id])
+        authorize encounter
         encounter.destroy
         render nothing: true, status: 204
       end
